Support setting files on cli
Using a local git dir is still supported but is now deprecated.

Instead just pass the files on the cli you want to release.

In addition, the validation logic was intertwined with run() and not
tested, so I pulled the parsing/validation of config and cli args into
the `config` module, and make it separately testable.