Fix path separator handling.
UrlFileName only does the separator conversion when part of a pydantic
model - my previous attempts to use it directly were just plain wrong.

The correct fix was to specify the name field to be the correct type,
and path separator normalisation works as intended, in that its now
impossible to send \ separated paths as part of the message.