Soft-disable releasing to github.
This change makes all releases go to job server by default.

If the workspace is still configured to release to github by default, we
warn the user, but release to job-server anyway.

The option is still there to release to github if tech-support team need
to, by adding the hidden --gh flag and setting PRIVATE_REPO_ACCESS_TOKEN
to the appropriate PAT in the env.

It cleans up some of the old local .git handling code as well.

Fixes #54