Add support for uploading to pre-existing release
This ended up being large, but was messy to pick apart.

The main change is that we now support `-r release_id` flag. If
specified, osrelease will not create a new release, and instead attempt
to upload to the existing release.

To do so, it needs to validate and display the list of files to upload
using the release-hatch api data for that release, rather than the local
file system. To achieve this, it needs calling the API in a new place in
the code, which got a bit messy. So this change splits out the config
loading from file list validation, to make that possible.

And since we now need test fixtures in multiple places, it improves the
text fixtures for creating workspace and releases